fix: extract duplicate components and clear stale selection state
- Extract HUD and PlaybackControls to src/components for reusability
- Add missing else clause in ArrowLODStream_2D/Map2D.jsx to clear stale selectedFeatures state
- Add missing else clause in ArrowLODStream_3D/Map3D.jsx to clear stale selectedFeatures state
- Update all four map implementations to import from shared components
- Remove duplicate component files from map subdirectories

This ensures proper state management when users clear their point selections,
preventing stale backend data from remaining visible in the UI.
