+++ src/maps/ArrowLODStream_2D/Map2D.jsx
@@ -4,8 +4,8 @@ import { TileLayer } from '@deck.gl/geo-layers';
 import { ScatterplotLayer, BitmapLayer } from '@deck.gl/layers';
 import { scaleLinear } from 'd3-scale';
 
-import { HUD } from './components/HUD';
-import { PlaybackControls } from './components/PlaybackControls';
+import { HUD } from '../../components/HUD';
+import { PlaybackControls } from '../../components/PlaybackControls';
 import { SelectionControls, DrawingOverlay, TimeSeriesChart, normalizeGeometry, filterPointsByGeometryInBounds } from '../../components/PointSelection';
 import ArrowLODTileLayer from '../../layers/ArrowLODTileLayer';
 
@@ -108,6 +108,8 @@ function ArrowLODStream2D() {
     useEffect(() => {
         if (selectedPointIds.size > 0) {
             queryBackendForSelection(selectedPointIds);
+        } else {
+            setSelectedFeatures(null);
         }
     }, [selectedPointIds]);
